+++ src/commands/stop.ts
@@ -1,22 +1,24 @@
 import Command from '../interfaces/command';
 import { Message } from 'discord.js';
 import bot from '..';
+import { MusicServer } from '../interfaces/music';
 
 export default class StopCommand implements Command {
     info = {
-        names: ['stop'],
+        names: ['leave', 'stop'],
         description: 'Stops playing',
-        usage: '&stop'
+        usage: '&stop',
+        category: 'music'
     }
 
     run(message: Message, _args: string[], messages: any): any {
         if(!message.member.voiceChannel)
             return message.reply(messages.connectVoice);
 
-        let server: any = bot.music[message.guild.id];
+        let server: MusicServer = bot.music[message.guild.id];
         if(server.dispatcher) {
-            delete bot.music[message.guild.id];
             message.member.voiceChannel.leave();
+            delete bot.music[message.guild.id];
             message.reply(messages.stopped);
         } else message.reply(messages.notPlaying);
     }